Add files via upload
This dataset uses 2010-2018 USA Census population and estimated population data and applies the equal proportions method. The goal of the analysis was to predict the number of House of Representatives members for each state expected from the 2020 census results which were not released at the time of this analysis. Results were accurate based on 2020 electoral votes.